+++ src/rs_driver/driver/input/unix/input_sock_select.hpp
@@ -66,7 +66,7 @@ private:
 
 protected:
   size_t pkt_buf_len_;
-  int fds_[2];
+  int fds_[3]{-1};
   size_t sock_offset_;
   size_t sock_tail_;
 };
@@ -78,7 +78,7 @@ inline bool InputSock::init()
     return true;
   }
 
-  int msop_fd = -1, difop_fd = -1;
+  int msop_fd = -1, difop_fd = -1, imu_fd = -1;
 
   msop_fd = createSocket(input_param_.msop_port, input_param_.host_address, input_param_.group_address);
   if (msop_fd < 0)
@@ -90,12 +90,23 @@ inline bool InputSock::init()
     if (difop_fd < 0)
       goto failDifop;
   }
-
   fds_[0] = msop_fd;
   fds_[1] = difop_fd;
 
+  if ((input_param_.imu_port != 0) && (input_param_.imu_port != input_param_.msop_port) &&  (input_param_.imu_port != input_param_.difop_port))
+  {
+    imu_fd = createSocket(input_param_.imu_port, input_param_.host_address, input_param_.group_address);
+    if (imu_fd < 0)
+      goto failImu;
+  }
+
+  fds_[2] = imu_fd;
+  
   init_flag_ = true;
   return true;
+
+failImu:
+  close(difop_fd);
 
 failDifop:
   close(msop_fd);
@@ -130,6 +141,9 @@ inline InputSock::~InputSock()
   close(fds_[0]);
   if (fds_[1] >= 0)
     close(fds_[1]);
+  
+  if(fds_[2] >= 0)
+    close(fds_[2]);
 }
 
 inline int InputSock::createSocket(uint16_t port, const std::string& hostIp, const std::string& grpIp)
@@ -232,7 +246,8 @@ failSocket:
 
 inline void InputSock::recvPacket()
 {
-  int max_fd = ((fds_[0] > fds_[1]) ? fds_[0] : fds_[1]);
+
+  int max_fd = std::max(std::max(fds_[0], fds_[1]), fds_[2]);
 
   while (!to_exit_recv_)
   {
@@ -242,6 +257,10 @@ inline void InputSock::recvPacket()
     if (fds_[1] >= 0)
       FD_SET(fds_[1], &rfds);
 
+    if(fds_[2] >= 0)
+    {
+      FD_SET(fds_[2], &rfds);
+    }
     struct timeval tv;
     tv.tv_sec = 1;
     tv.tv_usec = 0;
@@ -260,7 +279,7 @@ inline void InputSock::recvPacket()
       break;
     }
 
-    for (int i = 0; i < 2; i++)
+    for (int i = 0; i < 3; i++)
     {
       if ((fds_[i] >= 0) && FD_ISSET(fds_[i], &rfds))
       {