+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 # CHANGLOG
 
+# UnReleased
+
+
+## v1.5.17 2025-02-14
+
+### Added
+- Support RSAIRY.
+- Support parsing IMU data for RSAIRY and RSE1.
+- Support parsing IMU extrinsics parameters frome difop for RSAIRY.
+
+### Changed
+- Add feature attribute to point type.
+- Update help document.
+- Update block_time_offset as us for RSE1
+
+### Fix 
+- Fix compilation bug in win/input_socket_select.
+
+
 ## v1.5.16 2024-08-27
+
 ### Added
 - Support sn parsing of RSMX.
 ### Fixed
@@ -8,17 +28,22 @@
 - Fix compilation bug in epoll socket.
 - Fix compilation warning for unit tests.
 
+
 ## v1.5.15 2024-08-07
+
 ### Added
 - Support RSM3.
 
+
 ## v1.5.14 2024-07-15
+
 ### Added
 - Support multiple lidars with different multicast addresses and the same port.
 ### Fixed
 - Fix the bug that only one lidar was parsed correctly when multiple bp4.0 were used.
 
 ## v1.5.13 2024-05-10
+
 ### Added
 - Support RSMX.
 ### Fixed
@@ -26,10 +51,13 @@
 - Update firing_tss of Helios/Helios16P/RubyPlus.
 - Fix compilation bug of unit test.
 
+
 ## v1.5.12 2023-12-28
+
 ### Fixed
 - Fix bug in getting device info and status.
 - Fix bug in getting device temperature.
+
 
 ## v1.5.11 2023-12-18
 
@@ -43,6 +71,7 @@
 - Fix the bug of rs_drvier_viewer.cpp abnormal crash.
 - Fix fir_tss of RSBPV4.
 - Fix rotation direction of RSBPV4 when reversal.
+
 
 ## v1.5.10 2023-04-11